+++ cs25server/src/main/kotlin/edu/dongguk/cs25server/service/OrderService.kt
@@ -1,6 +1,5 @@
 package edu.dongguk.cs25server.service
 
-import edu.dongguk.cs25server.domain.Order
 import edu.dongguk.cs25server.domain.Store
 import edu.dongguk.cs25server.domain.type.OrderStatus
 import edu.dongguk.cs25server.dto.request.CustomerOrderRequestDto
@@ -10,7 +9,7 @@ import edu.dongguk.cs25server.dto.response.ListResponseDto
 import edu.dongguk.cs25server.dto.response.PageInfo
 import edu.dongguk.cs25server.exception.ErrorCode
 import edu.dongguk.cs25server.exception.GlobalException
-import edu.dongguk.cs25server.repository.OrderRepostiory
+import edu.dongguk.cs25server.repository.OrderRepository
 import edu.dongguk.cs25server.repository.StoreRepository
 import edu.dongguk.cs25server.util.Log.Companion.log
 import org.springframework.data.domain.Page
@@ -22,15 +21,18 @@ import org.springframework.transaction.annotation.Transactional
 
 @Service
 @Transactional
-class OrderService(private val storeRepository: StoreRepository, private val orderRepostiory: OrderRepostiory) {
+class OrderService(private val storeRepository: StoreRepository, private val orderRepository: OrderRepository) {
     // 점주 - 주문 내역 조회
     fun readCustomerOrder(storeId: Long, customerOrderRequestDto: CustomerOrderRequestDto): ListResponseDto<List<CustomerOrderResponseDto>> {
+        log.info("order_state: {}", customerOrderRequestDto.order_state)
+        log.info("start_date: {}", customerOrderRequestDto.start_date)
         val paging: Pageable = PageRequest.of(
                 customerOrderRequestDto.page.toInt(),
                 10,
         )
         val store: Store = storeRepository.findByIdOrNull(storeId) ?: throw GlobalException(ErrorCode.NOT_FOUND_STORE)
-        val customerOrderList: Page<OrderRepostiory.CustomerOrderInfo> = orderRepostiory.findCustomerOrderByStore(store, paging)
+        val customerOrderList: Page<OrderRepository.CustomerOrderInfo> =
+                orderRepository.findCustomerOrderByStore(store, OrderStatus.getCategory(customerOrderRequestDto.order_state), customerOrderRequestDto.start_date, customerOrderRequestDto.end_date, paging)
 
         val pageInfo: PageInfo = PageInfo(
                 page = customerOrderRequestDto.page.toInt(),
@@ -58,7 +60,7 @@ class OrderService(private val storeRepository: StoreRepository, private val ord
     // 점주 - 픽업 완료 처리
     fun pickupCustomerOrder(storeId: Long, customerPickupDtos: List<CustomerPickupRequestDto>): Boolean {
         customerPickupDtos.map { dto ->
-            orderRepostiory.findByIdOrNull(dto.customer_order_id)
+            orderRepository.findByIdOrNull(dto.customer_order_id)
                     ?.updateOrderStatus(OrderStatus.PICKUP)
                     ?: throw GlobalException(ErrorCode.NOT_FOUND_ORDER)
         }
